+++ Sources/Sentry/PrivateSentrySDKOnly.mm
@@ -207,11 +207,10 @@ static BOOL _framesTrackingMeasurementHybridSDKMode = NO;
         startSystemTime, endSystemTime, traceId, [SentrySDK currentHub]);
 
     if (payload != nil) {
+        const auto thread = sentry::profiling::ThreadHandle::current();
         payload[@"platform"] = SentryPlatformName;
-        payload[@"transaction"] = @{
-            @"active_thread_id" :
-                [NSNumber numberWithLongLong:sentry::profiling::ThreadHandle::current()->tid()]
-        };
+        payload[@"transaction"] =
+            @{ @"active_thread_id" : (thread == nullptr) ? @0 : @(thread->tid()) };
     }
 
     return payload;